Propagate xattr from UfsStatus to Inode
This change aims to allow the calls to the UFS from the
FileSystemMaster to propagate into the inode tree. This will
allow developers in the future to leverage any xattr that are
returned from the UFS implementations in order to make
optimizations or other assumptions about the state of a node
past the standard set of fields.

Most of these changes are mechanical, it's just a matter of
finding exactly where to set the xattr and where not to.

For now, the only usecase for setting xattr is during sync or
loadMetadata calls. The implementation adds the xattr to the
create*Context so that the inode tree can utilize set the xattr
on inodes from within the inode tree class.

pr-link: Alluxio/alluxio#9230
change-id: cid-b328d0215d41201cab90d4894c26e0e2d894bc59