Refactor Alluxio EMR script
This refactor includes a number of improvements to the Alluxio EMR
bootstrap script.

- Users can optionally specify a remote `alluxio-site.properties` file
to add arbitrary properties without adding them when launching
through a local `emr` command or in the web UI (Fixes #9489)
- Dynamically sets memory to 1/3 of available system RAM on
workers. Previously was hard-coded to 20GB (Closes #9478)
- Supports an optional argument for the user to specify the
delimiter used when passing arbitrary properties on the command
line. Previously this was hard-coded to `;` by default. Users can
choose to continue using `;` by specifying an empty string `""` or
`";"`. Additionally, in place of `;`, the third argument of the script
now allows the user to specify delimiter of their choice.
(Closes #9477)

Other various refactoring and bash programming style has been
utilized to reduce the amount of copy-and-paste between
commands throughout the script. Structure has been updated to
utilize functions more effectively along with a `main` function.

pr-link: Alluxio/alluxio#9505
change-id: cid-482ee3f86a47178ea41224c53d080425097c2e6b