[DOCFIX] Add mkdir step to Alluxio building guide
fixes #9251

For a user new to Alluxio and follows the Building-Alluxio-From-Source
guide for the first time, the cloned repository doesn't have directory
`./underFSStorage`. `bin/alluxio validateEnv local` will complain this
directory doesn't exist and fail.

I think it makes sense to add this step in the guide.

pr-link: Alluxio/alluxio#9627
change-id: cid-dbe1add359c2c296f4377167a7c83cf3fcc8681b